Made changes so Ragdolls and Gibs are recorded when an entity selected for recording is killed/destroyed.
Added member variable m_bEntRecIsRecording to c_baseanimating, so an entity knows if its being recorded.
Added RecordedEntityRagdolled() to CzmqManager which is hooked into BecomeRagdollOnClient() in c_baseanimating.
Added OnParentRagdolled() to CzmqBaseSkeletal, which changes the CzmqBaseSkeletal's mh_parent_entity to the handle of the newly created ragdoll.
Added OnEntityGibbed() to CzmqManager which is hooked into PropBreakableCreateAll() in props_shared, checks if newly created gibs belong to a recorded entity and records those gibs if so.

Removed mch_bonenames_list from CzmqBaseSkeletal because its useless now.

Fixed a number of bugs, and cleaned code up in some areas.